refactor (post) : global 도메인 내 일부파일 경로 변경으로 인한 import 문 정리
