+++ src/main/java/com/backend/immilog/post/presentation/controller/CommentController.java
@@ -5,7 +5,6 @@ import com.backend.immilog.post.presentation.request.CommentUploadRequest;
 import com.backend.immilog.post.presentation.response.PostApiResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
