+++ src/main/java/com/backend/immilog/post/presentation/controller/InteractionController.java
@@ -3,7 +3,6 @@ package com.backend.immilog.post.presentation.controller;
 import com.backend.immilog.post.application.services.InteractionCreationService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
